+++ databento/common/types.py
@@ -17,30 +17,6 @@ from databento.common.validation import validate_file_write_path
 
 logger = logging.getLogger(__name__)
 
-DBNRecord = (
-    databento_dbn.BBOMsg
-    | databento_dbn.CBBOMsg
-    | databento_dbn.CMBP1Msg
-    | databento_dbn.MBOMsg
-    | databento_dbn.MBP1Msg
-    | databento_dbn.MBP10Msg
-    | databento_dbn.TradeMsg
-    | databento_dbn.OHLCVMsg
-    | databento_dbn.ImbalanceMsg
-    | databento_dbn.InstrumentDefMsg
-    | databento_dbn.InstrumentDefMsgV1
-    | databento_dbn.InstrumentDefMsgV2
-    | databento_dbn.StatMsg
-    | databento_dbn.StatMsgV1
-    | databento_dbn.StatusMsg
-    | databento_dbn.SymbolMappingMsg
-    | databento_dbn.SymbolMappingMsgV1
-    | databento_dbn.SystemMsg
-    | databento_dbn.SystemMsgV1
-    | databento_dbn.ErrorMsg
-    | databento_dbn.ErrorMsgV1
-)
-
 _T = TypeVar("_T")
 
 
@@ -95,7 +71,7 @@ class MappingIntervalDict(TypedDict):
     symbol: str
 
 
-RecordCallback = Callable[[DBNRecord], None]
+RecordCallback = Callable[[databento_dbn.DBNRecord], None]
 ExceptionCallback = Callable[[Exception], None]
 ReconnectCallback = Callable[[pd.Timestamp, pd.Timestamp], None]
 
@@ -243,7 +219,7 @@ class ClientRecordCallback:
     def exc_callback_name(self) -> str:
         return getattr(self._exc_fn, "__name__", str(self._exc_fn))
 
-    def call(self, record: DBNRecord) -> None:
+    def call(self, record: databento_dbn.DBNRecord) -> None:
         """
         Execute the callback function, passing `record` in as the first
         argument. Any exceptions encountered will be dispatched to the