+++ databento/live/client.py
@@ -14,6 +14,7 @@ from typing import IO
 
 import databento_dbn
 import pandas as pd
+from databento_dbn import DBNRecord
 from databento_dbn import Schema
 from databento_dbn import SType
 
@@ -25,7 +26,6 @@ from databento.common.parsing import optional_datetime_to_unix_nanoseconds
 from databento.common.publishers import Dataset
 from databento.common.types import ClientRecordCallback
 from databento.common.types import ClientStream
-from databento.common.types import DBNRecord
 from databento.common.types import ExceptionCallback
 from databento.common.types import ReconnectCallback
 from databento.common.types import RecordCallback
@@ -57,7 +57,8 @@ class Live:
         gateway.
     heartbeat_interval_s: int, optional
         The interval in seconds at which the gateway will send heartbeat records if no
-        other data records are sent.
+        other data records are sent. By default heartbeats will be sent at the gateway's
+        default interval. Minimum interval is 5 seconds.
     reconnect_policy: ReconnectPolicy | str, optional
         The reconnect policy for the live session.
             - "none": the client will not reconnect (default)